fix reconnect, add delay after enter pressed after connect pressed

+++ src/main/java/com/RAmutex/utils/GlobalParameters.java
@@ -15,7 +15,7 @@ public class GlobalParameters
     public static int maxSectionOccupationTime = 5000;
     public static final String MAX_SECTION_OCCUPATION_TIME_DESCRIPTION = "max time of section occupation";
 
-    public static int getMaxSectionOccupationTimeWithDelay()
+    public static long getMaxSectionOccupationTimeWithDelay()
     {
 	return maxSectionOccupationTime + 1000;
     }
